# internal/docs/swagger.yaml
basePath: /
definitions:
...
  api.errorBody:
    properties:
      error:
        type: string
    type: object
...
  api.preferenceRequest:
    properties:
      autoDeleteAttachments:
        type: boolean
      dateFormat:
        type: string
      locale:
        type: string
      reviewPeriod:
        type: integer
      theme:
        type: string
      timeZone:
        type: string
      weekStart:
        type: integer
    type: object
...
  domain.Attachment:
    properties:
      contentType:
        type: string
      createdAt:
        type: string
      fileName:
        type: string
      id:
        type: integer
      size:
        type: integer
      todoId:
        type: integer
    type: object
  domain.AttachmentWithTodo:
    properties:
      contentType:
        type: string
      createdAt:
        type: string
      fileName:
        type: string
      id:
        type: integer
      size:
        type: integer
      todoDescription:
        type: string
      todoId:
        type: integer
      todoState:
        type: string
    type: object
...
  domain.Preference:
    properties:
      autoDeleteAttachments:
        description: |-
          AutoDeleteAttachments is a pointer because this column was added to a
          table that already had rows: schema sync adds new columns nullable with
          no default (see addMissingColumns), so a plain bool would fail to scan
          an existing row's NULL. Nil reads as off, same as false.
        type: boolean
      dateFormat:
        type: string
      locale:
        type: string
      reviewPeriod:
        type: integer
      theme:
        description: light | dark | system
        type: string
      timeZone:
        type: string
      updatedAt:
        type: string
      weekStart:
        description: 0 = Sunday
        type: integer
    type: object
...
  service.ContextCount:
    properties:
      contextId:
        type: integer
      name:
        type: string
      open:
        type: integer
    type: object
  service.MonthCount:
    properties:
      count:
        type: integer
      month:
        description: YYYY-MM
        type: string
    type: object
...
  service.Stats:
    properties:
      active:
        type: integer
      avgCompletionDays:
        type: number
      completed:
        type: integer
      completedLast30:
        type: integer
      completedLast365:
        type: integer
      completionRate:
        type: number
      deferred:
        type: integer
      oldestOpenDays:
        type: integer
      perContext:
        items:
          $ref: '#/definitions/service.ContextCount'
        type: array
      perMonth:
        items:
          $ref: '#/definitions/service.MonthCount'
        type: array
      projectsActive:
        type: integer
      projectsCompleted:
        type: integer
      projectsHidden:
        type: integer
      totalActions:
        type: integer
    type: object
...
info:
  contact: {}
  description: 'A GTD web application: actions, contexts, projects and the tickler.'
  title: gotracks API
  version: "1.0"
paths:
...
  /api/v1/attachments:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AttachmentWithTodo'
            type: array
      security:
      - BearerAuth: []
      summary: List all attachments
      tags:
      - attachments
  /api/v1/attachments/{id}:
    delete:
      parameters:
      - description: Attachment id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: deleted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Delete an attachment
      tags:
      - attachments
    get:
      parameters:
      - description: Attachment id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Download an attachment
      tags:
      - attachments
...
  /api/v1/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Export my data
      tags:
      - transfer
...
  /api/v1/preferences:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Preference'
      security:
      - BearerAuth: []
      summary: Get preferences
      tags:
      - preferences
    put:
      parameters:
      - description: Preferences
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.preferenceRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Preference'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Update preferences
      tags:
      - preferences
...
  /api/v1/stats:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Stats'
      security:
      - BearerAuth: []
      summary: Get statistics
      tags:
      - stats
...
  /api/v1/todos/{id}/attachments:
    get:
      parameters:
      - description: Action id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Attachment'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: List an action's attachments
      tags:
      - attachments
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Action id
        in: path
        name: id
        required: true
        type: integer
      - description: File to attach
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Attachment'
        "413":
          description: file too large
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Upload an attachment
      tags:
      - attachments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"